+++ src/service/test/constants.rs
@@ -1,0 +1,5 @@
+pub const TEST_DB_PREFIX: &str = "service-test-";
+pub const TEST_USERNAME: &str = "test_user";
+pub const TEST_PASSWORD: &str = "test_password";
+pub const TEST_MOUNT_NAME: &str = "collection";
+pub const TEST_MOUNT_SOURCE: &str = "test-data/small-collection";
